bot/requirements.txt
python-telegram-bot[job-queue]==20.7
httpx==0.27.0
python-dotenv==1.0.1
python-json-logger==2.0.7
pydantic==2.6.1
